Added comments for UnlockDoor script
Added XML documentation for each variable and method appearing in the UnlockDoor script. This script is the main puzzle piece for this project.

A console in the first room has a picture of a chess piece (a knight) with instructions to hold the chess piece over the console. Doing so will unlock the door. Once the player holds this chess piece over the console, the door is unlocked which is noted by a sound that gives feedback.