Added comments and cleaned up HighlightObject script
Added XML documentation to the variables and methods in the HighlightObject script.

This script detects when a player is in range of an interactable object. If they are in range and they aim a controller at the object, the object will change color to give feedback to the player that alerts them that they're able to pick up the object.